Mirror scikit api

+++ src/metrics/cluster_hcv.rs
@@ -37,5 +37,5 @@ macro_rules! cluster_hcv_struct {
   };
 }
 
-cluster_hcv_struct! {u32, Uint32Array}
+cluster_hcv_struct! {i64, BigInt64Array}
 cluster_hcv_struct! {u64, BigUint64Array}
